localRegistrationId should be int

+++ AxolotlKit/Classes/CipherMessage/PreKeyWhisperMessage.h
@@ -13,7 +13,7 @@
 
 - (instancetype)initWithData:(NSData *)serialized;
 
-- (instancetype)initWithWhisperMessage:(WhisperMessage*)whisperMessage registrationId:(long)registrationId prekeyId:(int)prekeyId signedPrekeyId:(int)signedPrekeyId baseKey:(NSData*)baseKey identityKey:(NSData*)identityKey;
+- (instancetype)initWithWhisperMessage:(WhisperMessage*)whisperMessage registrationId:(int)registrationId prekeyId:(int)prekeyId signedPrekeyId:(int)signedPrekeyId baseKey:(NSData*)baseKey identityKey:(NSData*)identityKey;
 
 @property (nonatomic, readonly) int            version;
 @property (nonatomic, readonly) NSData         *senderRatchetKey;
@@ -21,7 +21,7 @@
 @property (nonatomic, readonly) int            counter;
 @property (nonatomic, readonly) NSData         *cipherText;
 @property (nonatomic, readonly) NSData         *serialized;
-@property (nonatomic, readonly) long           registrationId;
+@property (nonatomic, readonly) int            registrationId;
 @property (nonatomic, readonly) int            prekeyID;
 @property (nonatomic, readonly) int            signedPrekeyId;
 @property (nonatomic, readonly) NSData         *baseKey;
