+++ AxolotlKit/Classes/Prekeys/PreKeyBundle.h
@@ -11,7 +11,7 @@
 @interface PreKeyBundle : NSObject
 
 @property (nonatomic, readonly) NSData *identityKey;
-@property (nonatomic, readonly) long   registrationId;
+@property (nonatomic, readonly) int    registrationId;
 @property (nonatomic, readonly) int    deviceId;
 @property (nonatomic, readonly) NSData *signedPreKeyPublic;
 @property (nonatomic, readonly) NSData *preKeyPublic;
@@ -19,7 +19,7 @@
 @property (nonatomic, readonly) int    signedPreKeyId;
 @property (nonatomic, readonly) NSData *signedPreKeySignature;
 
-- (instancetype)initWithRegistrationId:(long)registrationId
+- (instancetype)initWithRegistrationId:(int)registrationId
                               deviceId:(int)deviceId
                               preKeyId:(int)preKeyId
                           preKeyPublic:(NSData*)preKeyPublic
